COM-49 Added service classes and test methods

+++ src/main/java/com/teammental/common/dal/entity/City.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Entity
 @Table(name = "city")
-public class City extends BaseEntity<Short> {
+public class City extends BaseEntity<Integer> {
 
   @Column(name = "name")
   @NotNull
